% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFunctions.R
\name{EProjSimplexdiag}
\alias{EProjSimplexdiag}
\title{Solve the problem: min 1/2*x'*L*x-x'*d s.t. x>=0, 1'x=1}
\usage{
EProjSimplexdiag(d, l)
}
\arguments{
\item{d}{matrix or vector}

\item{l}{matrix or vector}
}
\value{
x
}
\description{
Solve the problem: min 1/2*x'*L*x-x'*d s.t. x>=0, 1'x=1
}
