% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFunctions.R
\name{EstClusters}
\alias{EstClusters}
\title{Estimate the optimal number of clusters C for clustering}
\usage{
EstClusters(drData, S)
}
\arguments{
\item{drData}{gene expression matrix after PCA processing}

\item{S}{the calculated similarity matrix S from "SimS"}
}
\value{
C the estimated number of clusters
}
\description{
Estimate the optimal number of clusters C for clustering
}
\examples{
gfData <- GenesFilter(yan)
res_SimS <- SimS(gfData)
EstClusters(res_SimS$drData,res_SimS$S)

}
