% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFunctions.R
\name{GenesFilter}
\alias{GenesFilter}
\title{Perform the step of gene filtering to normalizaed gene expression data}
\usage{
GenesFilter(data, gfRatio = 0.025)
}
\arguments{
\item{data}{the normalized gene expression matrix}

\item{gfRatio}{the ratio of genes filtering}
}
\value{
the gene expression matrix after genes filtering gfData
}
\description{
Perform the step of gene filtering to normalizaed gene expression data
}
\examples{
data(yan)
GenesFilter(yan)

}
