% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{PlotTrajectory}
\alias{PlotTrajectory}
\title{Infer the developmental trajectories based on the clustering results from RCSL}
\usage{
PlotTrajectory(
  gfData,
  clustRes,
  lineage,
  TrueLabel,
  fontSize = 12,
  dataName = "",
  VisualMethod = "umap"
)
}
\arguments{
\item{gfData}{preprocessed gene expression data (each column represent a cell)}

\item{clustRes}{the clustering results identified by RCSL}

\item{lineage}{the lineage obtained by getLineage()}
}
\value{
TrajectoryPlot ggplot object of the inferred developmental trajectories
}
\description{
Infer the developmental trajectories based on the clustering results from RCSL
}
