% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCSL.R
\name{RCSL}
\alias{RCSL}
\title{Perform the RCSL program}
\usage{
RCSL(
  data,
  GF = TRUE,
  gfRatio = 0.025,
  pcRatio = 0.95,
  NN.method = "KNN",
  Dis.method = "Euclidean",
  neiRatio = 0.65
)
}
\arguments{
\item{GF}{should I need the gene filter step?}

\item{gfRatio}{the ratio of the gene filter}

\item{pcRatio}{the ratio between the variance of the}

\item{NN.method}{the method of finding neighbors}

\item{Dis.method}{the distance metric in finding neighbors}

\item{neiRatio}{ratio of the number of selected}

\item{datat}{normalizaed gene expression matrix(each column represents a cell)}
}
\value{
gfData gene expression matrix after genes filtering

B block-diagonal matrix

C estimated number of clusters

y clustering results
}
\description{
Perform the RCSL program
}
\examples{
data(yan)
data <- log2(yan+1)
RCSL(yan[,1:20])
}
