% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFunctions.R
\name{SimS}
\alias{SimS}
\title{Calculate the initial similarity matrix}
\usage{
SimS(data, pcRatio = 0.95, largeThre = 2000, neiRatio = 0.65)
}
\arguments{
\item{data}{gene expression matrix after genes filtering (each column represents a cell)}

\item{pcRatio}{the ratio between the variance of the choosed PCs and the total variance}

\item{largeThre}{the number of cells to determine whether it is a large datasets}

\item{neiRatio}{ratio of the number of selected neighbors to the total number of cells}
}
\value{
S initial similarity matrix

drData gene expression matrix after PCA processing
}
\description{
Calculate the initial similarity matrix
}
