% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{TrajectoryAnalysis}
\alias{TrajectoryAnalysis}
\title{Trajectory analysis}
\usage{
TrajectoryAnalysis(
  gfData,
  drData,
  S,
  clustRes,
  fontSize = 12,
  TrueLabel,
  startPoint,
  dataName = "",
  sim = TRUE,
  simMeasure = "kendall",
  VisualMethod = "umap"
)
}
\arguments{
\item{gfData}{preprocessed gene expression data (each column represent a cell)}

\item{clustRes}{the clustering results identified by RCSL}

\item{lineage}{the lineage obtained by getLineage()}
}
\value{
PseudoTimePlot, MSTPlot, TrajectoryPlot
}
\description{
Trajectory analysis
}
\examples{
gfData <- GenesFilter(yan)
TrueLabel <- ann$cell_type1
res_SimS <- SimS(gfData)
C <- EstClusters(res_SimS$drData,res_SimS$S)
res_BDSM <- BDSM(res_SimS$S,C)
TrajectoryAnalysis(gfData,res_SimS$drData,res_SimS$S,res_BDSM$y,
                   TrueLabel=TrueLabel,startPoint=1)

}
