% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFunctions.R
\docType{data}
\name{yan}
\alias{yan}
\title{A public scRNA-seq dataset by Yan et al.}
\format{
An object of class \code{data.frame} with 20214 rows and 90 columns.
}
\source{
\url{http://dx.doi.org/10.1038/nsmb.2660}

Columns represent cells, rows represent genes expression values.
}
\usage{
yan
}
\description{
A public scRNA-seq dataset by Yan et al.
}
\keyword{datasets}
