% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFunctions.R
\name{GenesFilter}
\alias{GenesFilter}
\title{Perform the gene filter}
\usage{
GenesFilter(data, gfRatio = 0.025)
}
\arguments{
\item{data}{the normalized gene expression matrix (each column represents a cell)}

\item{gfRatio}{the ratio of genes filtering}
}
\value{
gfData gene expression matrix after genes filtering
}
\description{
Perform the gene filter
}
