% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFunctions.R
\name{NeigRepresent}
\alias{NeigRepresent}
\title{Calculate the neighbor representation of cells}
\usage{
NeigRepresent(drData, neiRatio = 0.65)
}
\arguments{
\item{drData}{gene expression matrix after dimensionality reduced by PCA}

\item{neiRatio}{ratio of the number of selected neighbors to the total number of cells}
}
\value{
NR the similarity matrix measured by neighbor representation
}
\description{
Calculate the neighbor representation of cells
}
