% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{PlotMST}
\alias{PlotMST}
\title{Plot the visualization of constructed Minimum Spanning Tree based on the clustering results of RCSL}
\usage{
PlotMST(
  drData,
  clustRes,
  TrueLabel,
  dataName = "",
  fontSize = 12,
  VisualMethod = "umap"
)
}
\arguments{
\item{drData}{preprocessed gene expression data (each column represent a cell)}

\item{clustRes}{the clustering results identified by RCSL}

\item{TrueLabel}{the real cell types to color the dots in plot}

\item{dataName}{plot title}

\item{fontSize}{the font size of the plot}

\item{VisualMethod}{the method for 2D visualization including UMAP,t-SNE and PCA}
}
\value{
MSTPlot ggplot object of the visualization of constructed MST
}
\description{
Plot the visualization of constructed Minimum Spanning Tree based on the clustering results of RCSL
}
