% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{PlotPseudoTime}
\alias{PlotPseudoTime}
\title{Infer the pseudo-temporal ordering between the cell types using the
distance from a cell type to the predefined starting cell type.}
\usage{
PlotPseudoTime(
  S,
  TrueLabel,
  startPoint,
  fontSize = 12,
  dataName = "",
  sim = TRUE
)
}
\arguments{
\item{S}{the similarity matrix calculated by SimS() function}

\item{TrueLabel}{the real cell types used to indicate the vertical axis}

\item{startPoint}{the posiition of the starting cell}

\item{fontSize}{the font size of the plot}
}
\value{
PstudoTime

PseudoTimePlot  ggplot object of the pseudo-temporal ordering of cells
}
\description{
Infer the pseudo-temporal ordering between the cell types using the
distance from a cell type to the predefined starting cell type.
}
