% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCSL.R
\name{RCSL}
\alias{RCSL}
\title{Perform the RCSL program}
\usage{
RCSL(
  data,
  GF = TRUE,
  gfRatio = 0.025,
  pcRatio = 0.95,
  largeThre = 2000,
  neiRatio = 0.65
)
}
\arguments{
\item{GF}{should I need the gene filter step?}

\item{gfRatio}{the ratio of the gene filter}

\item{pcRatio}{the ratio between the variance of the choosed PCs and the total variance}

\item{largeThre}{the number of cells to determine whether it is a large datasets}

\item{neiRatio}{ratio of the number of selected neighbors to the total number of cells in NR}

\item{datat}{normalizaed gene expression matrix(each column represents a cell)}
}
\value{
gfData gene expression matrix after genes filtering

B block-diagonal matrix

C estimated number of clusters

y clustering results
}
\description{
Perform the RCSL program
}
